LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/events/018-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo_raw' failed: Received invalid WebSocket response from the server

PASS open, message, error and close events
